miniproject/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
numpy==1.21.0
pickle-mixin==1.0.2
python-dotenv==0.19.0
google-generativeai